cr adjustments
